Properly bind link tooltips
Make sure it triggers only once and works with old-style RichTextField.
See https://github.com/springload/wagtail/commit/d8cf125d666c9799e069416da1f5c1fc6e14b309#commitcomment-17309996 for more information.
